+++ sdk/communication/communication-network-traversal/recordings/node/communicationnetworkingclient_playbacklive/recording_successfully_gets_a_turn_credential_providing_a_user_identity_and_any_routetype.js
@@ -1,0 +1,61 @@
+let nock = require('nock');
+
+module.exports.hash = "7aa8526889863bdfedf3f78d945ff207";
+
+module.exports.testInfo = {"uniqueName":{},"newDate":{}}
+
+nock('https://endpoint', {"encodedQueryParams":true})
+  .post('/identities')
+  .query(true)
+  .reply(201, {"identity":{"id":"sanitized"}}, [
+  'Content-Length',
+  '101',
+  'Content-Type',
+  'application/json; charset=utf-8',
+  'Request-Context',
+  'appId=',
+  'MS-CV',
+  '1lR66akHdk+pbuuraMiRfw.0',
+  'Strict-Transport-Security',
+  'max-age=2592000',
+  'x-ms-client-request-id',
+  'sanitized',
+  'api-supported-versions',
+  '2020-07-20-preview2, 2021-02-22-preview1, 2021-03-07, 2021-03-31-preview1, 2021-10-31-preview',
+  'X-Processing-Time',
+  '24ms',
+  'X-Cache',
+  'CONFIG_NOCACHE',
+  'X-Azure-Ref',
+  '03VV8YQAAAAD4e3M/OEmwRJisrhTyOG6cV1NURURHRTA4MTMAOWZjN2I1MTktYThjYy00Zjg5LTkzNWUtYzkxNDhhZTA5ZTgx',
+  'Date',
+  'Fri, 29 Oct 2021 20:13:16 GMT'
+]);
+
+nock('https://endpoint', {"encodedQueryParams":true})
+  .post('/networktraversal/:issueRelayConfiguration', {"id":"sanitized","routeType":"any"})
+  .query(true)
+  .reply(200, {"iceServers":[{"routeType":"any","urls":["turn.skype.com"],"username":"sanitized_username","credential":"sanitized_credential"}],"expiresOn":"2022-05-18T12:00:00.00+00:00"}, [
+  'Content-Length',
+  '319',
+  'Content-Type',
+  'application/json; charset=utf-8',
+  'Request-Context',
+  'appId=',
+  'MS-CV',
+  'oi+sP50CyEixd4MIzgJi3A.0',
+  'Strict-Transport-Security',
+  'max-age=2592000',
+  'x-ms-client-request-id',
+  'sanitized',
+  'api-supported-versions',
+  '2021-02-22-preview1, 2021-06-21-preview, 2021-10-08-preview',
+  'X-Processing-Time',
+  '23ms',
+  'X-Cache',
+  'CONFIG_NOCACHE',
+  'X-Azure-Ref',
+  '03VV8YQAAAAC21o1jyMehTbgSOsgBw0XNV1NURURHRTA4MDcAOWZjN2I1MTktYThjYy00Zjg5LTkzNWUtYzkxNDhhZTA5ZTgx',
+  'Date',
+  'Fri, 29 Oct 2021 20:13:16 GMT'
+]);