+++ sdk/communication/communication-network-traversal/recordings/node/communicationnetworkingclient_playbacklive/recording_successfully_gets_a_turn_credential_providing_a_user_identity_and_nearest_routetype.js
@@ -1,0 +1,61 @@
+let nock = require('nock');
+
+module.exports.hash = "34dbacd42f2bc3b8ef72f6c7b4504cd0";
+
+module.exports.testInfo = {"uniqueName":{},"newDate":{}}
+
+nock('https://endpoint', {"encodedQueryParams":true})
+  .post('/identities')
+  .query(true)
+  .reply(201, {"identity":{"id":"sanitized"}}, [
+  'Content-Length',
+  '101',
+  'Content-Type',
+  'application/json; charset=utf-8',
+  'Request-Context',
+  'appId=',
+  'MS-CV',
+  'sXnVsVPXjkyDagQHsrBf1Q.0',
+  'Strict-Transport-Security',
+  'max-age=2592000',
+  'x-ms-client-request-id',
+  'sanitized',
+  'api-supported-versions',
+  '2020-07-20-preview2, 2021-02-22-preview1, 2021-03-07, 2021-03-31-preview1, 2021-10-31-preview',
+  'X-Processing-Time',
+  '24ms',
+  'X-Cache',
+  'CONFIG_NOCACHE',
+  'X-Azure-Ref',
+  '03VV8YQAAAABGpMirKnvYTLfh+iHaAaYkV1NURURHRTA4MTMAOWZjN2I1MTktYThjYy00Zjg5LTkzNWUtYzkxNDhhZTA5ZTgx',
+  'Date',
+  'Fri, 29 Oct 2021 20:13:16 GMT'
+]);
+
+nock('https://endpoint', {"encodedQueryParams":true})
+  .post('/networktraversal/:issueRelayConfiguration', {"id":"sanitized","routeType":"nearest"})
+  .query(true)
+  .reply(200, {"iceServers":[{"routeType":"nearest","urls":["turn.skype.com"],"username":"sanitized_username","credential":"sanitized_credential"}],"expiresOn":"2022-05-18T12:00:00.00+00:00"}, [
+  'Content-Length',
+  '305',
+  'Content-Type',
+  'application/json; charset=utf-8',
+  'Request-Context',
+  'appId=',
+  'MS-CV',
+  'gjZC2PdEiEeTm1YRdZ7eRA.0',
+  'Strict-Transport-Security',
+  'max-age=2592000',
+  'x-ms-client-request-id',
+  'sanitized',
+  'api-supported-versions',
+  '2021-02-22-preview1, 2021-06-21-preview, 2021-10-08-preview',
+  'X-Processing-Time',
+  '24ms',
+  'X-Cache',
+  'CONFIG_NOCACHE',
+  'X-Azure-Ref',
+  '03VV8YQAAAADe8Nnsda1oRZ2xfNfdVUMwV1NURURHRTA4MDYAOWZjN2I1MTktYThjYy00Zjg5LTkzNWUtYzkxNDhhZTA5ZTgx',
+  'Date',
+  'Fri, 29 Oct 2021 20:13:17 GMT'
+]);