+++ sdk/communication/communication-network-traversal/recordings/node/communicationnetworkingclient_playbacklive/recording_successfully_gets_a_turn_credential_providing_a_user_identity_and_routetype.js
@@ -1,0 +1,61 @@
+let nock = require('nock');
+
+module.exports.hash = "f2a0ce94215787fc4a115835d626cee5";
+
+module.exports.testInfo = {"uniqueName":{},"newDate":{}}
+
+nock('https://endpoint', {"encodedQueryParams":true})
+  .post('/identities')
+  .query(true)
+  .reply(201, {"identity":{"id":"sanitized"}}, [
+  'Content-Length',
+  '101',
+  'Content-Type',
+  'application/json; charset=utf-8',
+  'Request-Context',
+  'appId=',
+  'MS-CV',
+  '5xRI94I1JkGJ9ij5XtiL5w.0',
+  'Strict-Transport-Security',
+  'max-age=2592000',
+  'x-ms-client-request-id',
+  'sanitized',
+  'api-supported-versions',
+  '2020-07-20-preview2, 2021-02-22-preview1, 2021-03-07, 2021-03-31-preview1, 2021-10-31-preview',
+  'X-Processing-Time',
+  '74ms',
+  'X-Cache',
+  'CONFIG_NOCACHE',
+  'X-Azure-Ref',
+  '0StttYQAAAAAyAVdGYR/ERqCE/Q1Di2yCV1NURURHRTA4MTAAOWZjN2I1MTktYThjYy00Zjg5LTkzNWUtYzkxNDhhZTA5ZTgx',
+  'Date',
+  'Mon, 18 Oct 2021 20:38:33 GMT'
+]);
+
+nock('https://endpoint', {"encodedQueryParams":true})
+  .post('/networktraversal/:issueRelayConfiguration', {"id":"sanitized","routeType":"any"})
+  .query(true)
+  .reply(200, {"iceServers":[{"routetype":"Any","urls":["turn.skype.com"],"username":"sanitized_username","credential":"sanitized_credential"}],"expiresOn":"2022-05-18T12:00:00.00+00:00"}, [
+  'Content-Length',
+  '319',
+  'Content-Type',
+  'application/json; charset=utf-8',
+  'Request-Context',
+  'appId=',
+  'MS-CV',
+  '+gxLHhirWkaVzMINVm9OBg.0',
+  'Strict-Transport-Security',
+  'max-age=2592000',
+  'x-ms-client-request-id',
+  'sanitized',
+  'api-supported-versions',
+  '2021-02-22-preview1, 2021-06-21-preview, 2021-10-08-preview',
+  'X-Processing-Time',
+  '433ms',
+  'X-Cache',
+  'CONFIG_NOCACHE',
+  'X-Azure-Ref',
+  '0StttYQAAAACLBDFRdF7yT5QVBvAsWg9qV1NURURHRTA4MjAAOWZjN2I1MTktYThjYy00Zjg5LTkzNWUtYzkxNDhhZTA5ZTgx',
+  'Date',
+  'Mon, 18 Oct 2021 20:38:34 GMT'
+]);