+++ sdk/communication/communication-network-traversal/src/communicationRelayClient.ts
@@ -26,6 +26,7 @@ import { createSpan } from "./common/tracing";
 import { CommunicationRelayClientOptions } from "./models";
 import {
   CommunicationRelayConfiguration,
+  CommunicationRelayConfigurationRequestRouteType,
   CommunicationNetworkTraversalIssueRelayConfigurationOptionalParams
 } from "./generated/src/models";
 
@@ -34,6 +35,7 @@ const isCommunicationRelayClientOptions = (
 ): options is CommunicationRelayClientOptions =>
   options && !isTokenCredential(options) && !isKeyCredential(options);
 
+interface GetRelayConfigurationOptions extends OperationOptions{}
 /**
  * Client class for the Azure Communication Services Relay client.
  */
@@ -121,15 +123,74 @@ export class CommunicationRelayClient {
   /**
    * Gets a TURN credential for a user
    *
-   * @param user - The user for whom to issue a token
    * @param options - Additional options for the request.
    */
   public async getRelayConfiguration(
-    user?: CommunicationUserIdentifier,
-    options: OperationOptions = {}
+    options?: GetRelayConfigurationOptions
+  ): Promise<CommunicationRelayConfiguration>;
+
+  /**
+   * Gets a TURN credential for a user
+   *
+   * @param routeType - The specified routeType for the relay request
+   * @param options - Additional options for the request.
+   */
+  public async getRelayConfiguration(
+    routeType: CommunicationRelayConfigurationRequestRouteType,
+    options?: GetRelayConfigurationOptions
+  ): Promise<CommunicationRelayConfiguration>;
+
+  /**
+   * Gets a TURN credential for a user
+   *
+   * @param user - The user for whom to issue a token
+   * @param routeType - The specified routeType for the relay request
+   * @param options - Additional options for the request.
+   */
+  public async getRelayConfiguration(
+    user: CommunicationUserIdentifier,
+    routeType?: CommunicationRelayConfigurationRequestRouteType,
+    options?: GetRelayConfigurationOptions
+  ): Promise<CommunicationRelayConfiguration>;
+
+  /**
+   * Gets a TURN credential for a user
+   *
+   * @param user - The user for whom to issue a token
+   * @param routeType - The specified routeType for the relay request
+   * @param options - Additional options for the request.
+   */
+  public async getRelayConfiguration(
+    paramOne?:
+      | CommunicationUserIdentifier
+      | CommunicationRelayConfigurationRequestRouteType
+      | GetRelayConfigurationOptions,
+    paramTwo?: CommunicationRelayConfigurationRequestRouteType | GetRelayConfigurationOptions,
+    options: GetRelayConfigurationOptions = {}
   ): Promise<CommunicationRelayConfiguration> {
-    const requestOptions: CommunicationNetworkTraversalIssueRelayConfigurationOptionalParams = options;
-    requestOptions.body = { id: user?.communicationUserId };
+    let requestOptions: CommunicationNetworkTraversalIssueRelayConfigurationOptionalParams = options;
+
+    if (
+      typeof paramOne !== "undefined" &&
+      typeof paramOne !== "string" &&
+      "communicationUserId" in paramOne
+    ) {
+      requestOptions.body = { id: paramOne.communicationUserId };
+      if (typeof paramTwo !== "undefined" && typeof paramTwo === "string") {
+        requestOptions.body["routeType"] = paramTwo;
+      }
+    } else if (typeof paramOne !== "undefined" && typeof paramOne === "string") {
+      requestOptions.body = { routeType: paramOne };
+      if (
+        typeof paramTwo !== "undefined" &&
+        typeof paramTwo !== "string" &&
+        "requestOptions" in paramTwo
+      ) {
+        requestOptions = paramTwo;
+      }
+    } else if (typeof paramOne !== "undefined" && "requestOptions" in paramOne) {
+      requestOptions = paramOne;
+    }
 
     const { span, updatedOptions } = createSpan(
       "CommunicationNetworkTraversal_IssueRelayConfiguration",