+++ sdk/communication/communication-network-traversal/src/index.ts
@@ -3,4 +3,9 @@
 
 export * from "./communicationRelayClient";
 export * from "./models";
-export { CommunicationIceServer, CommunicationRelayConfiguration } from "./generated/src/models";
+export {
+  CommunicationIceServer,
+  CommunicationRelayConfiguration,
+  CommunicationRelayConfigurationRequestRouteType,
+  CommunicationIceServerRouteType
+} from "./generated/src/models";