+++ sdk/communication/communication-network-traversal/test/public/communicationNetworkTraversalClient.spec.ts
@@ -45,12 +45,17 @@ matrix([[true, false]], async function(useAad) {
 
       const turnServers = turnCredentialResponse.iceServers;
 
+      // Should return both ANY and NEAREST routeType iceServers
+      assert.equal(turnServers.length, 2);
+
       for (const iceServer of turnServers) {
         for (const url of iceServer.urls) {
           assert.isNotNull(url);
         }
+
         assert.isNotNull(iceServer.username);
         assert.isNotNull(iceServer.credential);
+        assert.isNotNull(iceServer.routeType);
       }
     }).timeout(5000);
 
@@ -69,6 +74,55 @@ matrix([[true, false]], async function(useAad) {
         }
         assert.isNotNull(iceServer.username);
         assert.isNotNull(iceServer.credential);
+      }
+    }).timeout(5000);
+
+    it("successfully gets a turn credential providing a user identity and any routeType", async function() {
+      const connectionString = env.COMMUNICATION_LIVETEST_DYNAMIC_CONNECTION_STRING;
+      const identityClient = new CommunicationIdentityClient(connectionString);
+      const user: CommunicationUserIdentifier = await identityClient.createUser();
+
+      const turnCredentialResponse = await client.getRelayConfiguration(user, "any");
+      assert.isNotNull(turnCredentialResponse);
+
+      const turnTokenExpiresOn = turnCredentialResponse.expiresOn;
+      assert.isNotNull(turnTokenExpiresOn);
+
+      const turnServers = turnCredentialResponse.iceServers;
+
+      for (const iceServer of turnServers) {
+        for (const url of iceServer.urls) {
+          assert.isNotNull(url);
+        }
+
+        assert.isNotNull(iceServer.username);
+        assert.isNotNull(iceServer.credential);
+        assert.isNotNull(iceServer.routeType);
+        assert.equal(iceServer.routeType, "any");
+      }
+    }).timeout(5000);
+
+    it("successfully gets a turn credential providing a user identity and nearest routeType", async function() {
+      const connectionString = env.COMMUNICATION_LIVETEST_DYNAMIC_CONNECTION_STRING;
+      const identityClient = new CommunicationIdentityClient(connectionString);
+      const user: CommunicationUserIdentifier = await identityClient.createUser();
+
+      const turnCredentialResponse = await client.getRelayConfiguration(user, "nearest");
+      assert.isNotNull(turnCredentialResponse);
+
+      const turnTokenExpiresOn = turnCredentialResponse.expiresOn;
+      assert.isNotNull(turnTokenExpiresOn);
+
+      const turnServers = turnCredentialResponse.iceServers;
+
+      for (const iceServer of turnServers) {
+        for (const url of iceServer.urls) {
+          assert.isNotNull(url);
+        }
+        assert.isNotNull(iceServer.username);
+        assert.isNotNull(iceServer.credential);
+        assert.isNotNull(iceServer.routeType);
+        assert.equal(iceServer.routeType, "nearest");
       }
     }).timeout(5000);
   });